+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Core - Utilities
 
 - `isVisible` is added to `KeyboardObserver`
 - `reposition` method is added to `KeyboardObserver`
+- `value` property is added to `CachedManagedTask` and `CachedKeyedManagedTask`
 
 Extra - File Templates